fix: improve links styles; make more accessible
The biggest fix here is that the header link back to the home page now
has a `:focus` style, so when you tab through the links in the page you
can see that it is highlighted. Also, to minimize occurrences of this
style showing up as the result of mouse-based interaction, the icon no
longer links to the home page when you're already on the home page.

In addition, this removes the `scale` styles from the grid and the
buttons. The text in these elements did not scale well.